refactor: database entities

+++ server/src/main/kotlin/cz/globex/hana/core/dao/action/CreateDaoAction.kt
@@ -3,5 +3,5 @@ package cz.globex.hana.core.dao.action
 import cz.globex.hana.common.dto.*
 
 fun interface CreateDaoAction<T : EntityCreateDto> {
-	fun createOne(entity: T): Int
+	fun createOne(entity: T): Long
 }
