+++ server/src/main/kotlin/cz/globex/hana/core/dao/action/RetrieveDaoAction.kt
@@ -3,5 +3,5 @@ package cz.globex.hana.core.dao.action
 import cz.globex.hana.common.dto.*
 
 fun interface RetrieveDaoAction<R : EntityDto> {
-	fun retrieveOneOrNull(id: Int): R?
+	fun retrieveOneOrNull(id: Long): R?
 }
