+++ server/src/main/kotlin/cz/globex/hana/database/repository/ArticlesRepository.kt
@@ -3,6 +3,6 @@ package cz.globex.hana.database.repository
 import cz.globex.hana.database.entity.*
 import org.springframework.data.repository.*
 
-interface ArticlesRepository : PagingAndSortingRepository<Article, Int> {
-	fun getById(id: Int): Article?
+interface ArticlesRepository : PagingAndSortingRepository<Article, Long> {
+	fun getById(id: Long): Article?
 }
