+++ server/src/main/kotlin/cz/globex/hana/router/controller/action/DeleteAction.kt
@@ -6,5 +6,5 @@ import org.springframework.web.bind.annotation.*
 
 fun interface DeleteAction {
 	@DeleteMapping(path = ["/{${PathVariables.ID}}"])
-	fun deleteOne(@PathVariable(PathVariables.ID) id: Int): ResponseEntity<Unit>
+	fun deleteOne(@PathVariable(PathVariables.ID) id: Long): ResponseEntity<Unit>
 }
