+++ server/src/main/kotlin/cz/globex/hana/router/controller/action/RetrieveAction.kt
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.*
 
 fun interface RetrieveAction<R : EntityDto> {
 	@GetMapping(path = ["/{${PathVariables.ID}}"])
-	fun retrieveOne(@PathVariable(PathVariables.ID) id: Int): ResponseEntity<R>
+	fun retrieveOne(@PathVariable(PathVariables.ID) id: Long): ResponseEntity<R>
 
-	fun RetrieveDaoAction<R>.retrieveOneAndWrap(id: Int): ResponseEntity<R> {
+	fun RetrieveDaoAction<R>.retrieveOneAndWrap(id: Long): ResponseEntity<R> {
 		val entity = retrieveOneOrNull(id)
 		return if (entity != null) {
 			ResponseEntity.ok().body(entity)
