+++ server/src/main/kotlin/cz/globex/hana/router/controller/impl/EventsApiControllerImpl.kt
@@ -22,21 +22,21 @@ class EventsApiControllerImpl(daoProvider: DaoProvider) : EventsApiController {
 		return eventsDao.createOneAndWrap(entity)
 	}
 
-	override fun retrieveOne(id: Int): ResponseEntity<EventDto> = eventsDao.retrieveOneAndWrap(id)
+	override fun retrieveOne(id: Long): ResponseEntity<EventDto> = eventsDao.retrieveOneAndWrap(id)
 
-	override fun updateOne(id: Int, entity: EventCreateUpdateDto): ResponseEntity<Unit> {
+	override fun updateOne(id: Long, entity: EventCreateUpdateDto): ResponseEntity<Unit> {
 		TODO("Not yet implemented")
 	}
 
-	override fun deleteOne(id: Int): ResponseEntity<Unit> {
+	override fun deleteOne(id: Long): ResponseEntity<Unit> {
 		TODO("Not yet implemented")
 	}
 
-	override fun rateOne(id: Int, rate: RateDto): ResponseEntity<ResourceInfoDto> {
+	override fun rateOne(id: Long, rate: RateDto): ResponseEntity<ResourceInfoDto> {
 		TODO("Not yet implemented")
 	}
 
-	override fun reportOne(id: Int, report: ReportDto): ResponseEntity<ResourceInfoDto> {
+	override fun reportOne(id: Long, report: ReportDto): ResponseEntity<ResourceInfoDto> {
 		TODO("Not yet implemented")
 	}
 }
