+++ server/src/main/kotlin/cz/globex/hana/router/controller/impl/StockExchangesApiControllerImpl.kt
@@ -22,23 +22,23 @@ class StockExchangesApiControllerImpl(daoProvider: DaoProvider) : StockExchanges
 		return stockExchangesDao.createOneAndWrap(entity)
 	}
 
-	override fun retrieveOne(id: Int): ResponseEntity<StockExchangeDto> {
+	override fun retrieveOne(id: Long): ResponseEntity<StockExchangeDto> {
 		return stockExchangesDao.retrieveOneAndWrap(id)
 	}
 
-	override fun updateOne(id: Int, entity: StockExchangeCreateUpdateDto): ResponseEntity<Unit> {
+	override fun updateOne(id: Long, entity: StockExchangeCreateUpdateDto): ResponseEntity<Unit> {
 		TODO("Not yet implemented")
 	}
 
-	override fun deleteOne(id: Int): ResponseEntity<Unit> {
+	override fun deleteOne(id: Long): ResponseEntity<Unit> {
 		TODO("Not yet implemented")
 	}
 
-	override fun rateOne(id: Int, rate: RateDto): ResponseEntity<ResourceInfoDto> {
+	override fun rateOne(id: Long, rate: RateDto): ResponseEntity<ResourceInfoDto> {
 		TODO("Not yet implemented")
 	}
 
-	override fun reportOne(id: Int, report: ReportDto): ResponseEntity<ResourceInfoDto> {
+	override fun reportOne(id: Long, report: ReportDto): ResponseEntity<ResourceInfoDto> {
 		TODO("Not yet implemented")
 	}
 }
